### 2023-03-12 #5247 @editorialbot
[PRE REVIEW]: shoredate: An R package for shoreline dating coastal Stone Age sites in south-eastern Norway
**Submitting author:** @isakro (<a href="http://orcid.org/0000-0001-6974-1374">Isak Roalkvam</a>)
**Repository:** https://github.com/isakro/shoredate
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
