### 2023-03-12 #5247 @crvernon
:wave: @isakro while we wait on one more reviewer to confirm before we kick off the full review thread...could you go ahead and fix up these two missing DOIs in your bibliography?

> MISSING DOIs
> - 10.1111/j.1467-9876.2008.00623.x may be a valid DOI for title: A simple monotone process with application to radiocarbon-dated depth chronologies
> - 10.1017/rdc.2020.95 may be a valid DOI for title: Inference from large sets of radiocarbon dates: software and methods
