### 2023-03-12 #5247 @chrisleaman
Hi @crvernon, thanks for the ping! I'm not really an R programmer, so my review would be limited to the manuscript only. If this is okay, happy to take a look - else, I understand if you'd rather find someone else.
