### 2023-03-12 0016596
Updated DOIs references
