+++ paper/paper.bib
@@ -9,6 +9,7 @@
     number = {4},
     pages = {399--418},
     publisher = {Wiley Online Library},
+    doi = {10.1111/j.1467-9876.2008.00623.x},
     url =
       {https://rss.onlinelibrary.wiley.com/doi/full/10.1111/j.1467-9876.2008.00623.x},
   }
@@ -22,6 +23,7 @@
     volume = {63},
     issue = {1},
     pages = {23-39},
+    doi = {10.1017/rdc.2020.95},
     url = {https://dx.doi.org/10.1017/RDC.2020.95},
   }
 
