+++ NEWS.md
@@ -1,10 +1,10 @@
-# shoredate (development version)
+## shoredate 1.0.1.9000 (development version)
 
-## shoredate 1.0.1
+## shoredate 1.0.1 (2023-03-06)
 
  - `sum_shoredates()` can now handle NA dates.
  - `shoreline_date()` now accepts a vector of site names for the argument 
  `sites` if interpolated displacement curves are also provided.
 
-## shoredate 1.0.0 (19/01/2023)
+## shoredate 1.0.0 (2023-01-19)
 First release.
