+++ man/interpolate_curve.Rd
@@ -18,7 +18,7 @@ is interpolated.}
 
 \item{isobases}{4 spatial lines representing the shoreline isobases of the
 existing displacement curves. Multiple sets of 4 isobases with different
-isobase dirctions can be provided (see \code{\link[=create_isobases]{create_isobases()}}).}
+isobase directions can be provided (see \code{\link[=create_isobases]{create_isobases()}}).}
 
 \item{power}{A numerical value indicating the inverse distance power for IDW.
 Defaults to 2.}
@@ -45,12 +45,11 @@ the geologically derived displacement curves. This is done using inverse
 distance weighting.
 }
 \examples{
-\dontrun{
 # Create example point using the required coordinate system
 # WGS84 / zone UTM32N (EPSG: 32632)
 target_point <- sf::st_sfc(sf::st_point(c(579570, 6582982)), crs = 32632)
 
-# Interpolate shoreline displacement curve to the target point location.
-target_curve <- interpolate_curve(target_point)
-}
+# Interpolate shoreline displacement curve to the target point location,
+# setting the resolution on the calendar scale to 2000 years for speed.
+target_curve <- interpolate_curve(target_point, cal_reso = 2000)
 }
