+++ man/shoredate_sumplot.Rd
@@ -20,7 +20,6 @@ Function to plot the sum of the probabilities of multiple shoreline dates as
 resulting from running \code{sum_shoredates()}.
 }
 \examples{
-\dontrun{
 # Create example points
 target_points <- sf::st_sfc(sf::st_point(c(538310, 6544255)),
                             sf::st_point(c(538300, 6544250)))
@@ -28,13 +27,15 @@ target_points <- sf::st_sfc(sf::st_point(c(538310, 6544255)),
 # Set correct CRS
 target_points <- sf::st_as_sf(target_points, crs = 32632)
 
-# Perform shoreline dating
-target_dates <- shoreline_date(target_points, elevation = c(65, 70))
+# Reduce date resolution with cal_reso and elevation_reso for speed
+target_dates <- shoreline_date(target_points,
+                               elevation = c(65, 70),
+                               elev_reso = 10,
+                               cal_reso = 750)
 
 # Find summed probability
 target_sum <- sum_shoredates(target_dates)
 
 # Call to plot
 shoredate_sumplot(target_sum)
-}
 }
