+++ man/print.shoreline_date.Rd
@@ -19,15 +19,17 @@ HDRs contained in a \code{shoreline_date} object to console.
 Print the dates in a \code{shoreline_date} object. Each date is printed
 with site name, elevation and highest density region. If the isobase
 direction is different or more are provided than the default, the
-directions and dates associated with these are printed seperately.
+directions and dates associated with these are printed separately.
 }
 \examples{
-\dontrun{
 target_point <- sf::st_sfc(sf::st_point(c(538310, 6544255)), crs = 32632)
 
-target_date <- shoreline_date(site = target_point, elevation = 70)
+# Reduce date resolution with cal_reso and elevation_reso for speed
+target_date <- shoreline_date(site = target_point,
+                              elevation = 70,
+                              elev_reso = 1,
+                              cal_reso = 400)
 
 # Print to console
 target_date
-}
 }
