+++ man/sum_shoredates.Rd
@@ -14,11 +14,11 @@ sum_shoredates(
 \arguments{
 \item{shoreline_dates}{Object of class \code{shoreline_date}.}
 
-\item{cut_off}{Calender year specifying where dates should be cut off.
+\item{cut_off}{Calendar year specifying where dates should be cut off.
 Defaults to 2500 BCE.}
 
 \item{cut_off_level}{Numerical value between 0 and 1 indicating the
-probability mass that has to faller after the cut-off for a date to be
+probability mass that has to fall after the cut-off for a date to be
 excluded. Defaults to 1, retaining all dates.}
 
 \item{normalise}{Logical value indicating whether the probability sum of the
