+++ man/shoredate_plot.Rd
@@ -69,7 +69,7 @@ values denote CE.
 # Create example point with correct coordinate reference system
 target_point <- sf::st_sfc(sf::st_point(c(538310, 6544255)), crs = 32632)
 
-# Reduce date resolution with cal_reso and elevation_reso for speed.
+# Reduce date resolution with cal_reso and elevation_reso for speed
 target_date <- shoreline_date(sites = target_point, elevation = 80,
                               elev_reso = 10,
                               cal_reso = 500)
