### 2023-03-13 #5247 @crvernon
:wave: @TomasBeuzen are you interested in reviewing this submission for JOSS?  We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
