### 2023-03-28 #5247 @isakro
@crvernon  I took a look through the reviewer list and came across a few archaeologists who are proficient in R, in case you need suggestions for potential reviewers. For example benmarwick, joeroe and sebastien-plutniak should all be more than capable.
