### 2023-04-03 24e6045
shoredate_hdr() now returns weighted mean of shoreline date
