+++ R/shoreline_date.R
@@ -76,15 +76,15 @@
 #' @import terra
 #' @importFrom utils txtProgressBar
 #'
-#' @examples
-#' # Create example point using the required CRS WGS84 UTM32N (EPSG: 32632)
-#' target_point <- sf::st_sfc(sf::st_point(c(538310, 6544255)), crs = 32632)
-#'
 #' @references
 #' Roalkvam, I. 2023. A simulation-based assessment of the relation between
 #'  Stone Age sites and relative sea-level change along the Norwegian Skagerrak
 #'  coast. \emph{Quaternary Science Reviews} 299:107880. DOI:
 #'  https://doi.org/10.1016/j.quascirev.2022.107880
+#'
+#' @examples
+#' # Create example point using the required CRS WGS84 UTM32N (EPSG: 32632)
+#' target_point <- sf::st_sfc(sf::st_point(c(538310, 6544255)), crs = 32632)
 #'
 #' # Date target point, manually specifying the elevation instead of providing
 #' # an elevation raster. Reducing elev_reso and cal_reso for speed.
