+++ man/shoreline_date.Rd
@@ -103,6 +103,8 @@ subtracting the preceding value from each value of \code{px}.
 A function for shoreline dating Stone Age sites based on their present-day
 elevation, their likely elevation above sea-level when in use and the
 trajectory of past shoreline displacement on the Norwegian Skagerrak coast.
+Details and caveats pertaining to the implemented method is given in
+Roalkvam (2023).
 }
 \examples{
 # Create example point using the required CRS WGS84 UTM32N (EPSG: 32632)
@@ -114,4 +116,10 @@ shoreline_date(sites = target_point,
                elevation = 80,
                elev_reso = 1,
                cal_reso = 400)
+}
+\references{
+Roalkvam, I. 2023. A simulation-based assessment of the relation between
+Stone Age sites and relative sea-level change along the Norwegian Skagerrak
+coast. \emph{Quaternary Science Reviews} 299:107880. DOI:
+https://doi.org/10.1016/j.quascirev.2022.107880
 }
