+++ man/check_target_location.Rd
@@ -18,7 +18,7 @@ if it is lacking or is set to the incorrect coordinate reference system.
 Checks if a target site is set to the correct coordinate reference system
 WGS84 UTM32N (EPSG: 32632). Throws an error if this is not the case. Checks
 if the location intersects the study area for which the method was derived
-(see Roalkvam 2022). Returns a warning if this is not the case.
+(see Roalkvam 2023). Returns a warning if this is not the case.
 }
 \examples{
 # Create example point using the required coordinate system
@@ -28,4 +28,10 @@ target_point <- sf::st_sfc(sf::st_point(c(579570, 6582982)), crs = 32632)
 # Check the CRS and the location of the site
 check_target_location(target_point)
 
+}
+\references{
+Roalkvam, I. 2023. A simulation-based assessment of the relation between
+Stone Age sites and relative sea-level change along the Norwegian Skagerrak
+coast. \emph{Quaternary Science Reviews} 299:107880. DOI:
+https://doi.org/10.1016/j.quascirev.2022.107880
 }
