### 2023-04-03 80571d1
shoreline_date() returns weighted mean
