### 2023-04-04 a316eef
Added the paramter model to shoreline_date(), to allow for the assumption of no distance between site and shoreline.
