### 2023-04-04 #5247 @benmarwick
Yes, I can do this. I have reviewed JOSS papers previously and I'm familiar with the process
