### 2023-04-05 71cabd2
Skip test for target_plot() that fails with R CMD check. Minor edit to readme. Updated documentation.
