### 2023-04-05 #5337 @editorialbot
[REVIEW]: shoredate: An R package for shoreline dating coastal Stone Age sites in south-eastern Norway
**Submitting author:** @isakro (<a href="http://orcid.org/0000-0001-6974-1374">Isak Roalkvam</a>)
**Repository:** https://github.com/isakro/shoredate
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
