### 2023-04-06 738e82e
Updated shoredate_plot() for when shoreline_date() is run with model = "none". Removed dependency on ggsn for ggspatial in target_plot().
