+++ NAMESPACE
@@ -16,7 +16,7 @@ import(ggrepel)
 import(ggridges)
 import(sf)
 import(terra)
-importFrom(ggsn,scalebar)
+importFrom(ggspatial,annotation_scale)
 importFrom(grid,editGrob)
 importFrom(grid,gpar)
 importFrom(grid,grobHeight)
