### 2023-04-21 192b85a
Made target_plot() check for geom column
