### 2023-05-01 7f1c368
Recompiliation of documentation and news before release of v.1.0.2
