+++ NEWS.md
@@ -1,11 +1,10 @@
-## shoredate 1.0.1.9000 (development version)
+## shoredate 1.0.2 (2023-05-01)
  - `shoredate_hdr()` and `shoreline_date()` now returns the weighted mean date.
  - `shoreline_date()` now has the parameter model, which accepts either "none" 
  or "gamma". If "none" is specified then the distance between site and shoreline
  is assumed to be zero.
- - Now using ggspatial for scale bar in maps rather than ggsn, due to the 
- retiring of r-spatial packages 
- (see https://r-spatial.org/r/2022/04/12/evolution.html).
+ - Now using the package ggspatial for scale bar in maps rather than ggsn, due 
+ to the retiring of r-spatial packages (@rsbivand, #1).
 
 ## shoredate 1.0.1 (2023-03-06)
 
