### 2023-05-02 424e6c6
Recompiled target_plot() for paper. Updated date in NEWS.md. Resized first figure in readme
