### 2023-05-03 isakro/shoredate#2 @kanishkan91
JOSS review: error in target plot
I get an error when using the `target_plot()` function as mentioned in the README. The error is as below-

Error in `st_geometry<-.sf`(`*tmp*`, value = "geom") :
  inherits(x[[value]], "sfc") is not TRUE

https://github.com/openjournals/joss-reviews/issues/5337
