### 2023-05-03 isakro/shoredate#3 @kanishkan91
JOSS review: Remove some documentation from README and add github docs
The author has put together very detailed documentation. Can they consider generating github docs based on the README? All that would be required would be to run the following-

`library(pkgdown)`
`pkgdown::build_site()`

The docs can then be linked via a URL similar to this example- https://jgcri.github.io/ambrosia/ (The github repo for this example is here- https://github.com/JGCRI/ambrosia)

OPTIONAL- The README feels heavy. The author can put some of the documentation in a vignette. pkgdown will generate separate pages for different vignettes. I would be happy to describe how this can be implemented. It is quite simple.

https://github.com/openjournals/joss-reviews/issues/5337
