### 2023-05-03 isakro/shoredate#4 @kanishkan91
JOSS review: Remove scoped calls to functions. Leverage NAMESPACE in R instead.
There are several scoped calls to functions. For example in displacement_plot.R there are lines like `ggplot2::geom_line()`.

There should rather be lines in the function documentation like `from ggplot2 import geom_line` . Then when the NAMESPACE file is generated, all functions get documented in that file. See here for more details - https://bookdown.dongzhuoer.com/hadley/r-pkgs/namespace.html. I believe it is standard practice to not have scoped calls.

https://github.com/openjournals/joss-reviews/issues/5337
