### 2023-05-03 isakro/shoredate#5 @kanishkan91
JOSS review: Remove hard-coding of color schemes and line schemes
In some places there is hard coding of color schemes and line schemes- See line 69-78 here- https://github.com/isakro/shoredate/blob/master/R/displacement_plot.R

I think this reduces general usability of this software. For example, if a user wanted to add their own colors, it becomes unneccesarily hard.

Related to the above, I would recommend that these plot functions just return a plot object that can be plotted with any scheme.

So, for example this plotting would work like below-

`g <- displacement_plot(sample_data)`

Then, this g can be plotted with any user defined scheme- `g+scheme_basic`

https://github.com/openjournals/joss-reviews/issues/5337
