### 2023-05-03 #5337 @kanishkan91
I have almost completed a first pass of the review, I should be done soon (sorry it took so long). I think this is a great package that is very well documented and put together and I would be happy to recommend publication. However, there are a few things that I noted that the author can consider fixing (I have opened separate issues for these, but just summarizing here)-

1. I saw that the documentation is basically a very dense README file. The author can create github docs quite easily using the current documentation. I have described this in an issue for the author's convenience. Some of the documentation can also be moved to  a vignette or a separate page. Once again, this is described in the issue.
2. I observed that there are scoped calls to functions in many places in the codebase. For e.g.- `ggplot::geom_line()`  . I think the author can avoid using such scoped calls and instead leverage R's NAMESPACE capability here. I have described this once again in an issue.
3. I saw that color schemes and line schemes are hard coded within plotting functions. This can be parameterized so that users can add their own color and line schemes.
4. The target_plot() function failed for me. I followed the example from the README. Not sure why this happened.
