+++ NEWS.md
@@ -1,3 +1,5 @@
+# shoredate (development version)
+
 # shoredate 1.0.2
 
 ## shoredate 1.0.2 (2023-05-02)
