### 2023-05-03 isakro/shoredate#6 @benmarwick
JOSS review: add guidance for other researchers wanting to make a similar pkg for another region
I'm very pleased to review this high-quality software that I think is an excellent example for archaeologists to follow. Congratulations to the authors for their outstanding accomplishment.

My only suggestion for improvement is that based on the observation that currently the package is most useful for people interested in coastal Stone Age sites in south-eastern Norway, which I expect is a rather small group. One way to increase the value and impact of the package could be to include some instructions on how another researcher could make such a package for their own region of the world.

Perhaps this could be done in a new section in the existing vignette, or maybe a new brief vignette that, using the current package as an example, shares information about the data requirements and method steps that a researcher needs to know if they want to make a similar package for another part of the world.

https://github.com/openjournals/joss-reviews/issues/5337
