### 2023-05-04 #5337 @kanishkan91
@isakro Just going to add one more comment. I agree with @benmarwick that this software package seems very sptially specific (which is fine). But, its uses would increase tremendously if you added a vignette to just show users how they could use this for other regions. I think this would also warrant an additional paragraph in the paper. Perhaps after the addition of that vignette, you can even consider renaming this to something simler "An R package for shoreline dating coastal Stone Age sites" and just mention in the docs that currently this package is designed to do this for sites in South Eastern Norway, but can be extended to other sites (see vignette xxx.rmd). I think removing the hardcoding of names etc as I suggested in 1 issue would also help here.
