### 2023-05-09 ffc6e13
Added additional vignette on how to extend shoredate to other regions. Updated main vignette to reflect this.
