+++ vignettes/extending-shoredate.Rmd
@@ -1,0 +1,247 @@
+---
+title: "Applying shoredate to other regions"
+author: "Isak Roalkvam"
+output:
+  rmarkdown::html_vignette:
+    number_sections: true
+    toc: true
+vignette: >
+  %\VignetteIndexEntry{extending-shoredate}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+
+
+
+# Introduction
+
+While *shoredate* and the method underlying it has been developed for application in a constrained region of south-eastern Norway, it is possible to apply it to other areas. The following presents an example from Ørland in Trøndelag, Central Norway, for which a new displacement curve for the Late Holocene has recently been published (Romundset and Lakeman 2019). Furthermore, all that is needed to use the main functions of the package is a shoreline displacement curve that describes the sea-level change at the location of a site to be dated, and an eleveation value for that site. Consequently, it is in principle possible to extend the main functionality of the package to any region of the world in which the method is applicable, provided reliable data in a suitable format is available.
+
+Some notes of caution should also be added here. As the package was originally developed to be applied along the Skaggerrak coast of south-eastern Norway, the code associated with this kind of extension is not fully developed and is liable to be unstable. Furthermore, the functions `interpolate_curve()` and `create_isobases()` involve an increasing degree of uncertain extrapolation the further away from the spatial limit in south-eastern Norway one moves, and both functions will fail with different coordinate reference systems than WGS 82 / UTM zone 32N (EPSG:32632).
+
+Finally, the default method used for shoreline dating in the package is based on an empirically derived estimation of the relationship between coastal Stone Age sites and the contemporaneous shoreline in the Skaggrak region of south-eastern Norway. Consequently, the extent to which the same relationship characterises the site-sea relationship in other areas of Norway, and beyond, is therefore undetermined. However, given an assumption that when a location emerged from the sea determines the earliest possible date for when it could have been occupied, another option is to use the method to find a *terminus post quem* date, as outlined in the main vignette, and further demonstrated below.
+
+# Applying *shoredate* in Ørland
+
+## Shoreline displacement curve for Ørland
+
+Geologically derived displacement curves are in Norway often reported with a uniform probability between the lower and upper limit of the possible elevation of the sea-level over time. This is also the case for the displacement curve published for Ørland. At present, the required format for displacement curves to be used with *shoredate* is therefore a list of one or more displacement curves, where each curve is represented as a data frame with the columns `bce`, `upperelev` and `lowerelev`. Here, upperelev and lowerelev denote the highest and lowest possible elevation for the sea-level for each year BCE. Years BCE should here be given at an interval with equal or higher resolution than that which is desired for the resulting shoreline date. The probability between the upper and lower elevation limits are here assumed to be uniformly distributed. While displacement curves with a non-uniform probability distribution have also been reported for Norway (e.g. Creel et al. 2022), and are frequently developed for other regions of the world, functionality to use these with *shoredate* remains to be developed.
+
+The displacement curve for Ørland is provided with *shoredate* and can be loaded with the following code:
+
+
+```r
+# As a lot of function calls from these packages will be done below they are explictly loaded here,
+# instead of using the pacakge::function() syntax with each function call.
+library(ggplot2)
+library(sf)
+
+# Load shoredate
+# library(shoredate)
+
+# Load the displacement curve for Ørland
[…]
