### 2023-05-09 82cd369
Updated plotting functions with more parameters for graphical options. Adjusted shoreline_date() and target_plot() to accommodate application beyond SE Norway. Removed check_target_location() and handle the tests within the functions themselves.
