+++ tests/testthat/test-interpolate_curve.R
@@ -17,8 +17,8 @@ test_that("wrong CRS throws error and that this is printed", {
 test_that("if a site is located outside the limit of the study area, a warning is given", {
   skip_on_cran()
   target_point <- sf::st_sfc(sf::st_point(c(458310, 6544255)), crs = 32632)
-  err <- expect_warning(interpolate_curve(target_point))
-  expect_equal(err$message, "Target location is not within the study area for which the method was derived.")
+  warn <- expect_warning(interpolate_curve(target_point))
+  expect_equal(warn$message, "Target location is not within the study area for which the interpolation method was derived.")
 })
 
 test_that("the IDW works for a point on an isobase", {
