### 2023-05-09 76c9d5c
Reduced README and added references to vignettes. Updated documentation and news with new additions and adjustments.
