+++ README.Rmd
@@ -28,7 +28,7 @@ spatial_limit <- sf::st_read(system.file("extdata/spatial_limit.gpkg",
 [![Codecov test coverage](https://codecov.io/gh/isakro/shoredate/branch/master/graph/badge.svg)](https://app.codecov.io/gh/isakro/shoredate?branch=master)
 <!-- badges: end -->
 
-The package *shoredate* offers methods to shoreline date Stone Age sites located along the Norwegian Skagerrak coast based on their present-day elevation and the trajectory of past relative sea-level change. Shoreline dating is based on the premise that coastal Stone Age sites in the region were located on or close to the shoreline when they were in use, and is implemented here based on an empirically derived estimate of the likely elevation of the sites above sea-level when they were occupied (Roalkvam 2023). However, do note that as the method is dependent on regularities in human behaviour and as the Roalkvam (2023) study provides an initial formalisation of the method, it is hefted with unexplored uncertainties. Consequently, the dates achieved with the package should be treated with care.  
+The package *shoredate* offers methods to shoreline date coastal Stone Age sites based on their present-day elevation and the trajectory of past relative sea-level change. Shoreline dating is based on the premise that coastal Stone Age sites in large parts of Fennoscandia were located on or close to the shoreline when they were in use. The package and method as implemented here was originally developed for the Norwegian Skagerrak coast in south-eastern Norway, based on an empirically derived estimate of the likely elevation of sites above sea-level when they were occupied (Roalkvam 2023). However, do note that as the method is dependent on regularities in human behaviour, and as the Roalkvam (2023) study provides an initial formalisation of the method, it is hefted with unexplored uncertainties. Consequently, the dates achieved with the package should be treated with care.  
 
 ## Installation and loading
 
@@ -53,9 +53,10 @@ library(shoredate)
 
 ## Geographical and temporal coverage
 
-As the method of shoreline dating is determined by relative sea-level change, it is dependent on reliable geological reconstructions of this development. At present, the method as outlined here is therefore limited to being applicable in the region of south-eastern Norway between Horten in the north east to Arendal in the south west. This region has newly compiled shoreline displacement curves for Horten (Romundset 2021), Porsgrunn (Sørensen et al. 2014; Sørensen et al. 2023), Tvedestrand (Romundset 2018; Romundset et al. 2018) and Arendal (Romundset 2018). The region also formed the study area for Roalkvam (2023), in which the method and its parameters were derived. The spatial coverage is indicated in the maps below. The shoreline isobases in the second figure represent contours along which the shoreline displacement has followed the same trajectory. These correspond to the displacement curves and place names in the third figure, which also indicates the temporal coverage of the package. 
+As the method of shoreline dating is determined by relative sea-level change, it is dependent on reliable geological reconstructions of this development.  The method as outlined here was therefore originally developed for the Skagerrak region of south-eastern Norway -- extending from Horten in the north east to Arendal in the south west. This region has newly compiled shoreline displacement curves for Horten (Romundset 2021), Porsgrunn (Sørensen et al. 2014; Sørensen et al. 2023), Tvedestrand (Romundset 2018; Romundset et al. 2018) and Arendal (Romundset 2018). The region also formed the study area for Roalkvam (2023), in which the method and its parameters were derived. The remainder of this document and the [main vignette](doc/shoredate.html) focuses on this area. It is, however, possible to adapt the package to be applicable in other regions (as outlined in the [second vignette](doc/extending-shoredate.html).
 
-Note that spatial data used with the package should be set to WGS 84 / UTM zone 32N (EPSG:32632).
+The location of the 
[…]
