+++ DESCRIPTION
@@ -30,8 +30,6 @@ Imports:
     ggrepel,
     ggridges,
     ggspatial,
-    grid,
-    scales,
     sf,
     terra,
     utils
@@ -42,6 +40,7 @@ Suggests:
     knitr,
     rmarkdown,
     testthat (>= 3.0.0),
-    vdiffr
+    vdiffr,
+    rnaturalearth
 VignetteBuilder: knitr
 Config/testthat/edition: 3
