+++ NAMESPACE
@@ -1,7 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
 S3method(print,shoreline_date)
-export(check_target_location)
 export(create_isobases)
 export(displacement_plot)
 export(interpolate_curve)
@@ -11,19 +10,4 @@ export(shoredate_sumplot)
 export(shoreline_date)
 export(sum_shoredates)
 export(target_plot)
-import(ggplot2)
-import(ggrepel)
-import(ggridges)
-import(sf)
-import(terra)
-importFrom(ggspatial,annotation_scale)
-importFrom(grid,editGrob)
-importFrom(grid,gpar)
-importFrom(grid,grobHeight)
-importFrom(grid,grobTree)
-importFrom(grid,grobWidth)
-importFrom(grid,textGrob)
-importFrom(grid,unit.c)
-importFrom(grid,viewport)
-importFrom(scales,squish_infinite)
-importFrom(utils,txtProgressBar)
+importFrom(ggplot2,.data)
