### 2023-05-09 e6f4c70
Updated second vignette
