+++ _pkgdown.yml
@@ -1,0 +1,4 @@
+url: https://isakro.github.io/shoredate/
+template:
+  bootstrap: 5
+
