+++ pkgdown.yml
@@ -1,0 +1,11 @@
+pandoc: 2.19.2
+pkgdown: 2.0.7
+pkgdown_sha: ~
+articles:
+  extending-shoredate: extending-shoredate.html
+  shoredate: shoredate.html
+last_built: 2023-05-09T22:17Z
+urls:
+  reference: https://isakro.github.io/shoredate/reference
+  article: https://isakro.github.io/shoredate/articles
+
