### 2023-05-09 7c26630
Added additional tests to target_plot()
