### 2023-05-10 75067ff
Changes to documentation. Updated functions that deal with displacement curves to accept these as both data frames and lists of data frames.
