+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   extending-shoredate: extending-shoredate.html
   shoredate: shoredate.html
-last_built: 2023-05-09T23:31Z
+last_built: 2023-05-10T19:28Z
 urls:
   reference: https://isakro.github.io/shoredate/reference
   article: https://isakro.github.io/shoredate/articles
