### 2023-05-11 940b1a0
Skipping rnaturalearth test due to issues with GitHub. Made the natrualearth_zoom argument of target_plot() take two values, seperating adjustmens on the x and y axes. Some adjustments to documentation.
