+++ R/target_plot.R
@@ -19,50 +19,52 @@
 #'  (EPSG:32632).
 #' @param naturalearth_basemap Logical value specifying if a background map
 #'  should be downloaded to be used as a basemap. Downloaded files are stored
-#'  with `base::tempdir()` and deleted when the R session is closed. If TRUE,
-#'  overrides the `basemap` argument. Defaults to FALSE.
-#' @param naturalearth_zoom A numerical value specifying the amount of cropping
-#'  that is done around provided `targets` when `naturalearth_basemap` is set to
-#'  TRUE. Be aware of whether a projected or geographical CRS is specified in
-#'  `crs_epsg`. Defaults to 20000.
+#'  with `base::tempdir()` and deleted when the R session is closed. If `TRUE`,
+#'  overrides the `basemap` argument. Defaults to `FALSE`.
+#' @param naturalearth_zoom A vector of two numerical values specifying the
+#'  amount of cropping that is done around provided `targets` when
+#'  `naturalearth_basemap` is set to
+#'  `TRUE`. Be aware of whether a projected or geographical CRS is specified in
+#'  `crs_epsg`. Defaults to `c(20000, 20000)`.
 #' @param target_labels Logical value specifying whether the targets should be
 #'  labelled in the plot. Takes the first column beyond the one holding the
 #'  geometries to represent names. If this is not present the targets are
-#'  labelled by row number. Defaults to TRUE.
+#'  labelled by row number. Defaults to `TRUE`.
 #' @param scalebar Logical specifying whether a scalebar should be added to the
-#'  plot. Defaults to TRUE.
+#'  plot. Defaults to `TRUE`.
 #' @param scalebar_width Numerical value specifying the width of the scalebar by
 #'  passing it to the `width_hint` argument of `ggspatial::annotation:scale()`.
-#'  Defaults to 0.4.
+#'  Defaults to `0.4`.
 #' @param scalebar_style Character value specifying the style of the scalebar by
 #'  passing it to the `style` argument of `ggspatial::annotation:scale()`.
-#'  Defaults to "ticks".
+#'  Defaults to `"ticks"`.
 #' @param scalebar_location Character value specifying the location of the
 #'  scalebar on the plot by passing it to the `location` argument of
-#'  `ggspatial::annotation:scale()`. Defaults to "br".
+#'  `ggspatial::annotation:scale()`. Defaults to `"br"`.
 #' @param base_fill Character value specifying the fill colour of the basemap.
-#'  Defaults to "grey".
+#'  Defaults to `"grey"`.
 #' @param base_col Character value specifying the outline colour of the
-#'  basemap. Defaults to NA.
+#'  basemap. Defaults to `NA`.
 #' @param target_shape Numerical value specifying the point shape that represent
-#'  the centroids of the targets.
+#'  the centroids of the targets. Defaults to `21`.
 #' @param target_col Character value specifying the colour parameter for the
-#'  points that represent the centroids of the targets.
+#'  points that represent the centroids of the targets. Defaults to `"black"`.
 #' @param target_fill Character value specifying the fill parameter for the
-#'  points that represent the centroids of the targets.
+#'  points that represent the centroids of the targets. Defaults to `"red"`.
 #' @param target_size Numerical value specifying the size of the points that
-#'  represent the centroids of the targets.
+#'  represent the centroids of the targets. Defaults to `2.25`.
 #' @param isobase_line Vector of character values specifying the linetype that
 #'  is used to represent the isobases of the geologically derived displacement
[…]
