+++ vignettes/extending-shoredate.Rmd
@@ -26,13 +26,13 @@ Finally, the default method used for shoreline dating in the package is based on
 
 ## Shoreline displacement curve for Ørland
 
-Geologically derived displacement curves are in Norway often reported with a uniform probability between the lower and upper limit of the possible elevation of the sea-level over time. This is also the case for the displacement curve published for Ørland. At present, the required format for displacement curves to be used with *shoredate* is therefore a data frame with the columns `bce`, `upperelev` and `lowerelev`. Here, `upperelev` and `lowerelev` denote the highest and lowest possible elevation for the sea-level for each year BCE. Years BCE, defined by the column `bce`, should be given at an interval with equal or higher resolution than that which is desired for the resulting shoreline date. The probability between the upper and lower elevation limits are here assumed to be uniformly distributed. While displacement curves with a non-uniform probability distribution have also been reported for Norway (e.g. Creel et al. 2022), and are frequently developed for other regions of the world, functionality to use these with *shoredate* remains to be developed.
+Geologically derived displacement curves are in Norway often reported with a uniform probability between the lower and upper limit of the possible elevation of the sea-level over time. This is also the case for the displacement curve published for Ørland. At present, the required format for displacement curves to be used with *shoredate* is therefore a data frame with the columns `bce`, `upperelev`, `lowerelev` and `name`. Here, `upperelev` and `lowerelev` denote the highest and lowest possible elevation for the sea-level for each year BCE. Years BCE, defined by the column `bce`, should be given at an interval with equal or higher resolution than that which is desired for the resulting shoreline date. The probability between the upper and lower elevation limits are here assumed to be uniformly distributed. While displacement curves with a non-uniform probability distribution have also been reported for Norway (e.g. Creel et al. 2022), and are frequently developed for other regions of the world, functionality to use these with *shoredate* remains to be developed.
 
 The displacement curve for Ørland is provided with *shoredate* and can be loaded with the following code:
 
 
 ```r
-# As a lot of function calls from these packages will be done below they are
+# As functions from these packages will be used quite a bit below they are
 # explictly loaded here, instead of using the package::function() syntax with
 # each function call.
 library(ggplot2)
@@ -44,19 +44,22 @@ library(sf)
 # Load the displacement curve for Ørland
 orland_disp <- get(load(system.file("extdata/orland_displacement_curve.rda",
                   package = "shoredate")))
+
+# Print the first five rows of the Ørland displacement curve
+head(orland_dip)
 ```
 
-This can then be plotted with displacement_plot():
+```
+## Error in head(orland_dip): object 'orland_dip' not found
+```
+
+This can then be plotted with `displacement_plot()`. The name is adjusted with the argument `target_name` to something more informative, and the geologically derived displacement curves from south-eastern Norway are excluded by specifying the `displacement_curves` argument, as these are not of relevance here.
 
 
 ```r
-# As the displacement curves for south-eastern Norway are not of interest here,
-# these are not plotted by specifying the argument displacement_curves.
-# Functions from ggplot2 are then applied to perform some adjustments to the
-# plot returned by displacement_plot().
-displacement_plot(target_curve = orland_disp, displacement_curves = "") +
-            theme(legend.position = "none") +
-            ggtitle("Shoreline displacement curve for Ørland, Central Norway")
+displacement_plot(target_curve = orland_disp,
+                  target_name = "Displacement curve for Ørland, Central Norway",
+                  displacement_curves = NA)
 ```
 
 <img src="extend-orland-disp-1.png" alt="plot of chunk orland-disp" style="display: block; margin: auto;" />
[…]
