### 2023-05-11 45eb4f4
Fixed some spelling errors and updated documentation.
