### 2023-05-11 049a5af
Reworded DESCRIPTION, vignette and README. Adjusted the coverage figures used in README, vignette and paper.
