+++ vignettes/extending-shoredate.Rmd
@@ -20,7 +20,7 @@ While *shoredate* and the method underlying it has been developed for applicatio
 
 Some notes of caution should also be added here. As the package was originally developed to be applied along the Skaggerrak coast of south-eastern Norway, the code associated with this kind of extension is not fully developed and is liable to be unstable. Furthermore, the function `interpolate_curve()` involves an increasing degree of uncertain extrapolation the further away from the spatial limit in south-eastern Norway one moves, and both `interpolate_curve()` and `create_isobases()` will fail with other coordinate reference systems than WGS 82 / UTM zone 32N (EPSG:32632).
 
-Finally, the default method used for shoreline dating in the package is based on an empirically derived estimation of the relationship between coastal Stone Age sites and the contemporaneous shoreline in the Skagerrak region of south-eastern Norway. Consequently, the extent to which the same relationship characterises the site-sea relationship in other areas of Norway, and beyond, is therefore undetermined. However, given an assumption that when a location emerged from the sea determines the earliest possible date for when it could have been occupied, another option is to use the method to find a *terminus post quem* date, as outlined in the main vignette, and further demonstrated below.
+Finally, the default method used for shoreline dating in the package is based on an empirically derived estimation of the relationship between coastal Stone Age sites and the contemporaneous shoreline in the Skagerrak region of south-eastern Norway. Consequently, the extent to which the same relationship characterises the site-sea relationship in other areas of Norway, and beyond, is therefore undetermined (a possible approach for assessing this in other areas can be found in Roalkvam 2023). However, in regions characterised by continous relative sea-level regression, when a location emerged from the sea determines the earliest possible date for when it could have been occupied. Consequently, another option is to use the method to find a *terminus post quem* date, as outlined in the main vignette, and further demonstrated below.
 
 # Applying *shoredate* in Ørland
 
@@ -46,11 +46,17 @@ orland_disp <- get(load(system.file("extdata/orland_displacement_curve.rda",
                   package = "shoredate")))
 
 # Print the first five rows of the Ørland displacement curve
-head(orland_dip)
+head(orland_disp)
 ```
 
 ```
-## Error in head(orland_dip): object 'orland_dip' not found
+##    bce upperelev    lowerelev   name
+## 1 1950 0.3372681 0.0000000000 Ørland
+## 2 1949 0.3376883 0.0003724002 Ørland
+## 3 1948 0.3381084 0.0007448004 Ørland
+## 4 1947 0.3385286 0.0011172007 Ørland
+## 5 1946 0.3389487 0.0014896009 Ørland
+## 6 1945 0.3393688 0.0018620011 Ørland
 ```
 
 This can then be plotted with `displacement_plot()`. The name is adjusted with the argument `target_name` to something more informative, and the geologically derived displacement curves from south-eastern Norway are excluded by specifying the `displacement_curves` argument, as these are not of relevance here.
@@ -70,7 +76,7 @@ Furthermore, while variable uplift rates are also relevant for this area, this i
 
 ## Creating maps of Ørland
 
-Having loaded the displacement curve, this can then be directly used to shoreline date sites in locations where it applies, provided the elevation of the sites above present sea-level is known. First we will create a couple of fictitious site examples, each represented by a point, and maps of their location to demonstrate the extendability of some additional functions.
[…]
