+++ README.Rmd
@@ -30,7 +30,7 @@ spatial_limit <- sf::st_read(system.file("extdata/spatial_limit.gpkg",
 
 The package *shoredate* offers methods to shoreline date coastal Stone Age sites based on their present-day elevation and the trajectory of past relative sea-level change. Shoreline dating is based on the premise that coastal Stone Age sites in large parts of Fennoscandia were located on or close to the shoreline when they were in use. The package and method as implemented here was originally developed for the Norwegian Skagerrak coast in south-eastern Norway, based on an empirically derived estimate of the likely elevation of sites above sea-level when they were occupied (Roalkvam 2023). 
 
-While the package offers ways to extend and adjust the method for application in other areas, provided suitable data is available, the ways and degree to which the procedures are directly applicable elsewhere is largely undetermined and likely to vary between contexts. Do also note that as the method is dependent on regularities in human behaviour, and as the Roalkvam (2023) study provides an initial formalisation of the method, it is hefted with unexplored uncertainties, also within the core area for which it was developed. In sum therefore, the dates achieved with the package should be treated with care.  
+While the package offers ways to extend and adjust the method for application in other areas which have been characterised by a monotonic trajectory of relative sea-level regression, the ways and degree to which the procedures are directly applicable elsewhere is largely undetermined and likely to vary between contexts. Do also note that as the method is dependent on regularities in human behaviour, and as the Roalkvam (2023) study provides an initial formalisation of the method, it is hefted with unexplored uncertainties, also within the core area for which it was developed. In sum therefore, the dates achieved with the package should be treated with care.  
 
 ## Installation and loading
 
