+++ NEWS.md
@@ -10,12 +10,13 @@
  transferability of the package (@benmarwick, #6).
 - Made most graphical parameters in `shoredate_plot()`, `displacement_plot()` 
  and `target_plot()` variables that be adjusted by the user (@kanishkan91, #5).
-- Adjusted how the "geom" column of targets passed to `target_plot()` are 
+- Adjusted how the `geom` column of `sf` objects passed to `target_plot()` are 
  evaluated (@kanishkan91, #2).
-- Cleaned up code to reduce the amount of Import and ImportFrom in NAMESPACE,
- opting instead for the `package::function()` syntax (@kanishkan91, #4).
-- Set up a pkgdown site for the package through GitHub at 
-  isakro.github.io/shoredate/ (@kanishkan91, #3).
+- Cleaned up code to reduce the amount of `@Import` and `@ImportFrom` in 
+ NAMESPACE, opting instead for the `package::function()` syntax
+ (@kanishkan91, #4).
+- Set up a pkgdown site for the package, served through GitHub at 
+  https://isakro.github.io/shoredate/ (@kanishkan91, #3).
 
 ## shoredate 1.0.2 (2023-05-02)
  - `shoredate_hdr()` and `shoreline_date()` now returns the weighted mean date.
