### 2023-05-15 4f22822
displacement_plot() now consistently orders legend elements. As a result, also re-did figures and documentation
