+++ vignettes/extending-shoredate.Rmd
@@ -36,13 +36,7 @@ The displacement curve for Ørland is provided with *shoredate* and can be loade
 # each function call.
 library(ggplot2)
 library(sf)
-```
 
-```
-## Linking to GEOS 3.8.0, GDAL 3.0.4, PROJ 7.0.0; sf_use_s2() is TRUE
-```
-
-```r
 # Load shoredate
 library(shoredate)
 
