### 2023-05-15 isakro/shoredate#2 @isakro
This should now be resolved as I have changed way the `geom` column of a `sf` object is evaluated by `target_plot()`.
