### 2023-05-15 isakro/shoredate#3 @isakro
I have been wondering about the extent of the README as well. To make the README more succinct, and as it is already repeated in the main vignette, I have therefore removed the example of how to perform and handle multiple dates from the README.

Furthermore, as mentioned in #6, I have also added a second vignette that outlines how the package could be applied in other regions, which also demonstrates some further functionality. Finally, I have also made a `pkgdown` site hosted with GitHub Pages: isakro.github.io/shoredate/ (which was surprisingly painless to set up).
