### 2023-05-15 isakro/shoredate#4 @isakro
Thank you for pointing this out - I seem to have been a bit inconsistent and had not realised all the nuances to this. I have now read more on when to choose between `@Import`, `@ImportFrom`, and `pkg::fun()`, as well as looked more into the relationship between the DESCRIPTION and the NAMESPACE files. I think I have arrived on `package::function()` being the best option when calling external functions.

In the latest version of the R Packages book, Wickham and Bryan recommend defaulting to the `pkg::fun()` syntax:
https://r-pkgs.org/dependencies-in-practice.html#sec-dependencies-in-imports-r-code

And from 11.6 Imports in the link you provided:
"It’s common for packages to be listed in Imports in DESCRIPTION, but not in NAMESPACE. In fact, this is what I recommend: list the package in DESCRIPTION so that it’s installed, then always refer to it explicitly with pkg::fun(). Unless there is a strong reason not to, it’s better to be explicit."

For the above reasons, and unless you have any objections, I will therefore use the `pkg::fun()` syntax for external function calls, so as to avoid having to use `@Import` and `@ImportFrom`, and thus avoid having packages loaded through NAMESPACE. This is with the exception of `@ImportFrom ggplot2 .data`, as this appears to be necessary for it be possible to use the `.data` notation with the plotting functions.
