### 2023-05-15 isakro/shoredate#5 @isakro
Good points! I have now changed it so that most predefined graphical parameters in the plotting functions are instead the defaults of parameters that a user can manipulate.

Furthermore, `target_plot()`, `displacement_plot()` and `shoredate_sumplot()` all return `ggplot` objects that can be manipulated further (also exemplified in the second vignette). E.g.
```
g <- displacement_plot()
g + theme_void()
```

However, one exception to this is `shoredate_plot()`. Due to the many options for how the dating procedure can be performed, the plotting function returns a somewhat convoluted nested list of lists of plots which have to be unpacked in the correct way. I am yet to find a solution that handles this in a way that can return a series of `ggplot` objects instead of sequentially calling `print()` on these. However, if the adjustments that can be done via the plotting arguments of the function is OK for now, I will open this as a separate issue that I can try to deal with at a later stage - I fear it might take quite a bit of effort to resolve.
