### 2023-05-15 isakro/shoredate#6 @isakro
Thank you for the good suggestion, and for taking the time to review this.

I have now added the vignette ["Applying shoredate to other regions"](https://isakro.github.io/shoredate/articles/extending-shoredate.html) (along with some additional functionality). This demonstrates ways in which this could be done, using an example from another area of Norway, and includes some pointers for how others can explore these possibilities as well. While a lot still remains to be done in terms of applicability elsewhere, especially as the method is currently only applicable in regions with a continuous sea-level fall, it is something I also intend to work on going forward.
