### 2023-05-16 #5337 @kanishkan91
@isakro Last comment (I promise!). Based on latest additions, could you consider changing your install instructions to

`devtools::install_github('JGCRI/shoredate', build_vignettes = TRUE) `

This would automatically build vignettes for users. My apologies for the last minute comment.
