+++ paper/paper.bib
@@ -193,7 +193,7 @@
 	title = {{Holocen vegetasjonshistorie og landhevning i søndre Vestfold og sørøstre
 Telemark}},
 	author = {Sørensen, Rolf and Kari E Henningsmoen and Helge I Høeg and Veronika Gälman},
-	editor = {{Per}, {Persson} and {Steinar}, {Solheim}},
+	editor = {{Persson}, {Per}, and {Solheim}, {Steinar}},
 	booktitle = {{The Stone Age in Telemark. Archaeological Results and Scientific
 Analysis from Vestfoldbaneprosjektet and E18 Rugtvedt--Dørdal}},
 	year = {2023},
