+++ README.Rmd
@@ -30,7 +30,7 @@ spatial_limit <- sf::st_read(system.file("extdata/spatial_limit.gpkg",
 
 The package *shoredate* offers methods to shoreline date coastal Stone Age sites based on their present-day elevation and the trajectory of past relative sea-level change. Shoreline dating is based on the premise that coastal Stone Age sites in large parts of Fennoscandia were located on or close to the shoreline when they were in use. The package and method as implemented here was originally developed for the Norwegian Skagerrak coast in south-eastern Norway, based on an empirically derived estimate of the likely elevation of sites above sea-level when they were occupied (Roalkvam 2023). 
 
-While the package offers ways to extend and adjust the method for application in other areas, the ways and degree to which the procedures are directly applicable elsewhere is largely undetermined, and likely to vary between contexts. Such extensions are currently limited to regions that have been characterised by a monotonic trajectory of relative sea-level regression. Do also note that as the method is dependent on regularities in human behaviour, and as the Roalkvam (2023) study provides an initial formalisation of the method, it is hefted with unexplored uncertainties, also within the core area for which it was developed. In sum therefore, the dates achieved with the package should be treated with care.  
+While the package offers ways to extend and adjust the method for application in other areas, the ways and degree to which the procedures are directly applicable elsewhere is largely undetermined and likely to vary between contexts. Furthermore, such extensions are currently limited to regions that have been characterised by a monotonic trajectory of relative sea-level regression. Do also note that as the method is dependent on regularities in human behaviour, and as the Roalkvam (2023) study provides an initial formalisation of the method, it is hefted with unexplored uncertainties, also within the core area for which it was developed. In sum therefore, the dates achieved with the package should be treated with care.  
 
 ## Installation and loading
 
@@ -44,7 +44,7 @@ The latest development version can be installed from [GitHub](https://github.com
 
 ```{r setup, eval = FALSE}
 # install.packages("devtools")
-devtools::install_github("isakro/shoredate")
+devtools::install_github("isakro/shoredate", build_vignettes = TRUE)
 ```
 
 When it has been installed the package can be loaded:
@@ -55,7 +55,7 @@ library(shoredate)
 
 ## Geographical and temporal coverage
 
-As the method of shoreline dating is determined by relative sea-level change, it is dependent on reliable geological reconstructions of this development.  The method as outlined here was therefore originally developed for the Skagerrak region of south-eastern Norway -- extending from Horten in the north east to Arendal in the south west (see figure below). This region has newly compiled shoreline displacement curves for Horten (Romundset 2021), Porsgrunn (Sørensen et al. 2014; Sørensen et al. 2023), Tvedestrand (Romundset 2018; Romundset et al. 2018) and Arendal (Romundset 2018). The region also formed the study area for Roalkvam (2023), in which the method and its parameters were derived. The remainder of this document and the [main vignette](https://isakro.github.io/shoredate/articles/shoredate.html) focuses on this area. It is, however, possible to adapt the procedures for application in other regions, which is outlined in the [second vignette](https://isakro.github.io/shoredate/articles/extending-shoredate.html).
[…]
