+++ README.md
@@ -26,15 +26,15 @@ occupied (Roalkvam 2023).
 
 While the package offers ways to extend and adjust the method for
 application in other areas, the ways and degree to which the procedures
-are directly applicable elsewhere is largely undetermined, and likely to
-vary between contexts. Such extensions are currently limited to regions
-that have been characterised by a monotonic trajectory of relative
-sea-level regression. Do also note that as the method is dependent on
-regularities in human behaviour, and as the Roalkvam (2023) study
-provides an initial formalisation of the method, it is hefted with
-unexplored uncertainties, also within the core area for which it was
-developed. In sum therefore, the dates achieved with the package should
-be treated with care.
+are directly applicable elsewhere is largely undetermined and likely to
+vary between contexts. Furthermore, such extensions are currently
+limited to regions that have been characterised by a monotonic
+trajectory of relative sea-level regression. Do also note that as the
+method is dependent on regularities in human behaviour, and as the
+Roalkvam (2023) study provides an initial formalisation of the method,
+it is hefted with unexplored uncertainties, also within the core area
+for which it was developed. In sum therefore, the dates achieved with
+the package should be treated with care.
 
 ## Installation and loading
 
@@ -50,7 +50,7 @@ The latest development version can be installed from
 
 ``` r
 # install.packages("devtools")
-devtools::install_github("isakro/shoredate")
+devtools::install_github("isakro/shoredate", build_vignettes = TRUE)
 ```
 
 When it has been installed the package can be loaded:
@@ -63,11 +63,11 @@ library(shoredate)
 
 As the method of shoreline dating is determined by relative sea-level
 change, it is dependent on reliable geological reconstructions of this
-development. The method as outlined here was therefore originally
-developed for the Skagerrak region of south-eastern Norway – extending
-from Horten in the north east to Arendal in the south west (see figure
-below). This region has newly compiled shoreline displacement curves for
-Horten (Romundset 2021), Porsgrunn (Sørensen et al. 2014; Sørensen et
+development. The method as outlined here was originally developed for
+the Skagerrak region of south-eastern Norway – extending from Horten in
+the north east to Arendal in the south west (see figure below). This
+region has newly compiled shoreline displacement curves for Horten
+(Romundset 2021), Porsgrunn (Sørensen et al. 2014; Sørensen et
 al. 2023), Tvedestrand (Romundset 2018; Romundset et al. 2018) and
 Arendal (Romundset 2018). The region also formed the study area for
 Roalkvam (2023), in which the method and its parameters were derived.
@@ -88,18 +88,19 @@ temporal coverage of the method within the region.
 
 As human occupation only occurred some time after the retreat of the
 Fennoscandian Ice Sheet, the currently oldest known sites in Norway are
-from around 9300 BCE (e.g. Glørstad 2016). The oldest possible age to
-achieve with *shoredate* in the Skagerrak region of south-eastern Norway
-is 9469 BCE, although no sites are yet known to be that old. If a site
-has an elevation that implies a date older than the lower temporal limit
-of the displacement curves, it is returned as NA and a warning is given.
+from around 9300 BCE (e.g. Glørstad 2016). Although no sites are yet
+known to be that old, the earliest possible age to achieve with
+*shoredate* in the Skagerrak region of south-eastern Norway is 9469 BCE,
+which marks the latest start date among the employed displacement
+curves. If a site has an elevation that implies a date older than this,
+it is returned as NA and a warning is given.
 
 In Roalkvam (2023) it was found that sites in the study region tend to
 be located on or close to the shoreline up until around the transition
 to the Late Neolithic, c. 2500 BCE, which thus marks the upper limit for
[…]
