+++ vignettes/shoredate.Rmd
@@ -171,7 +171,7 @@ tpq_date <- shoreline_date(target_point,
                            model = "none")
 
 # Plot and adjust the position of the HDR label on the y-axis
-shoredate_plot(tpq_date, hdr_label_yadj = 0.5)
+shoredate_plot(tpq_date, hdr_label_yadj = 0.6)
 ```
 
 <img src="shore-tpq_date-1.png" alt="plot of chunk tpq_date" style="display: block; margin: auto;" />
