### 2023-05-19 #5337 @isakro
Thanks @kanishkan91, that's useful, I hadn't realised the vignettes aren't built automatically. I have updated the readme and vignette.
