### 2023-05-21 #5337 @crvernon
> @isakro I think naming it "shoredate: An R package for shoreline dating coastal Stone Age sites" is justified (to me atleast) given the additions. Also I think this name would encourage others to contribute to this software so that it can one day also track regions which are are subject to more complex situations. But would also be interested to see what the editor (@crvernon ) and the other reviewer (@benmarwick) say. Either way, great job addressing the comments! I'm happy to clear this for publication.

@kanishkan91 concerning the title, I agree that generalizing the package name to not be regionally specific would make the software more appealing to a larger audience.  Ultimately, I'll let this be up to @isakro
