### 2023-05-21 18c288f
Fixed Sørensen et al. 2023 reference
