### 2023-05-22 812b2f1
Minor edits to function documentation
