+++ R/shoredate_sumplot.R
@@ -5,7 +5,7 @@
 #'
 #' @param shoredates_sum Object of class `shoredates_sum`.
 #' @param sample_size Logical indicating whether or not to display the number of
-#'  summed dates on the plot. Defaults to TRUE.
+#'  summed dates on the plot. Defaults to `TRUE`.
 #'
 #' @return A line plot showing the provided summed probability distribution.
 #' @export
