+++ R/shoredate_plot.R
@@ -5,49 +5,52 @@
 #' @param shorelinedates Object of class `shoreline_date`.
 #' @param date_probability Logical value indicating whether the
 #'  probability distribution of the shoreline date should be plotted.
-#'  Defaults to TRUE.
+#'  Defaults to `TRUE`.
 #' @param elevation_distribution Logical value indicating whether the
 #'  distribution describing the distance between site and shoreline should be
-#'  displayed. Default is TRUE.
+#'  displayed. Default is `TRUE`.
 #' @param displacement_curve Logical value indicating whether the displacement
-#'  curve should be displayed. Default is TRUE.
+#'  curve should be displayed. Default is `TRUE`.
 #' @param site_name Logical value indicating whether the name of the site should
-#'  be printed in the header of the plot. Defaults to FALSE.
+#'  be printed in the header of the plot. Defaults to `FALSE`.
 #' @param parameters Logical value indicating whether the parameters of the
-#'  statistical function should be displayed. Default is FALSE.
+#'  statistical function should be displayed. Default is `FALSE`.
 #' @param isobase_direction  Logical value indicating whether the direction of
-#'  the isobases should be printed. Default is FALSE.
+#'  the isobases should be printed. Default is `FALSE`.
 #' @param highest_density_region Logical value indicating whether the 95%
-#'  highest density region should be displayed. Defaults to TRUE.
+#'  highest density region should be displayed. Defaults to `TRUE`.
 #' @param hdr_label Logical value indicating whether the numeric values for the
-#'  highest density regions should be displayed. Default is TRUE.
+#'  highest density regions should be displayed. Default is `TRUE`.
 #' @param multiplot Logical value indicating whether multiple dates should be
 #'  plotted individually, or be collapsed into a single plot. The only other
-#'  graphical option with `multiplot` set to TRUE is `highest_density_region`.
-#'  Default is FALSE.
+#'  graphical option with `multiplot` set to `TRUE` is `highest_density_region`.
+#'  Default is `FALSE`.
 #' @param date_col Character value specifying the outline colour of the
-#'  probability distribution of the shoreline date.
+#'  probability distribution of the shoreline date. Defaults to `NA`.
 #' @param date_fill Character value specifying the fill colour of the
-#'  probability distribution of the shoreline date.
+#'  probability distribution of the shoreline date. Defaults to `"darkgrey"`.
 #' @param displacement_col Character value specifying the outline colour of the
-#'  displacement curve.
+#'  displacement curve. Defaults to `"red"`.
 #' @param displacement_fill Character value specifying the fill colour of the
-#'  displacement curve.
+#'  displacement curve. Defaults to `"red"`.
 #' @param site_elevation_col Character value specifying the outline colour of
 #'  the distribution describing the likely distance between site and shoreline.
+#'  Defaults to `"#046c9a"`.
 #' @param site_elevation_fill Character value specifying the fill colour of
 #'  the distribution describing the likely distance between site and shoreline.
+#'  Defaults to `"#046c9a"`.
 #' @param hdr_col Character value specifying the colour of the line
-#'  segment giving the highest density region of the shoreline date.
+#'  segment giving the highest density region of the shoreline date. Defaults to
+#'   `"black"`.
 #' @param hdr_label_xadj Numerical value between 0 and 1 specifying the position
 #'  of the HDR label on the x-axis. Increasing the value moves the label further
-#'  from the plot border. Defaults to 0.2.
+#'  from the plot border. Defaults to `0.2`.
 #' @param hdr_label_yadj Numerical value between 0 and 1 specifying the position
 #'  of the HDR label on the y-axis. Increasing the value moves the label further
[…]
