### 2023-05-25 #5337 @crvernon
Very well done @isakro and reviewers!  You have produced a very nice paper and software product.

We are almost there! Next is just setting up the archive for your new release.

So here is what we have left to do:

- [ ] Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)

- [ ] Please respond with the DOI of the archived version here
