### 2023-05-25 #5337 @isakro
Great, thanks @crvernon!

I have made a GitHub release of [v1.1.0](https://github.com/isakro/shoredate/releases/tag/v1.1.0)
This has been archived on Zenodo with the DOI: [10.5281/zenodo.7971859](https://doi.org/10.5281/zenodo.7971859)
